+++ cmd/dougless/main.go
@@ -5,17 +5,24 @@ import (
 	"os"
 
 	"github.com/douglasjordan/dougless-runtime/internal/runtime"
+	"github.com/douglasjordan/dougless-runtime/internal/repl"
 )
 
 func main() {
-	if len(os.Args) < 2 {
-		fmt.Println("Usage: dougless <script.js>")
-		os.Exit(1)
-	}
-
-	scriptPath := os.Args[1]
-	
 	rt := runtime.New()
+
+  // go into repl mode if no args
+  if len(os.Args) < 2 {
+    r := repl.New(rt, os.Stdin, os.Stdout)
+    if err := r.Run(); err != nil {
+      fmt.Fprintf(os.Stderr, "REPL Error: %v\n", err)
+      os.Exit(1)
+    }
+    return
+  }
+
+  // the only other accepted arg are .js files
+	scriptPath := os.Args[1]
 	if err := rt.ExecuteFile(scriptPath); err != nil {
 		fmt.Fprintf(os.Stderr, "Error: %v\n", err)
 		os.Exit(1)